requirements.txt
# Python 3.10 compatibility
# Last updated: May 4, 2025
# Standardized Python version across all project components
# Core dependencies
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.5.0  # Updated for Python 3.10 compatibility
python-dotenv>=1.0.0
pandas>=2.2.0  # Updated for Python 3.10 compatibility
numpy>=1.26.0
scikit-learn>=1.4.0
streamlit>=1.32.0
pytest>=8.0.0
python-jose>=3.3.0
passlib>=1.7.4
argon2-cffi>=21.3.0
argon2-cffi-bindings==21.2.0  # Pinned version for Python 3.10 compatibility
cryptography>=42.0.0
httpx>=0.26.0
aiohttp>=3.9.0
spacy>=3.7.0
nltk>=3.8.1
transformers>=4.38.0
torch>=2.2.0
jax>=0.4.20  # Added JAX for high-performance machine learning
ray>=2.9.0  # Added Ray for distributed AI workloads

# MLOps and AI-specific libraries
mlflow>=2.10.0  # MLOps lifecycle management
huggingface_hub>=0.20.0  # Model hub integration
optuna>=3.5.0  # Hyperparameter optimization
wandb>=0.16.0  # Experiment tracking

# LLM and embeddings
openai>=1.10.0  # OpenAI API
langchain>=0.1.10  # For LLM workflows
langchain-openai>=0.0.5  # OpenAI integration
langchain-anthropic>=0.1.0  # Anthropic integration
langchain-google-genai>=0.0.5  # Google Gemini integration
langchain-community>=0.0.15  # Community integrations
tiktoken>=0.5.1  # Token counting for OpenAI models
tenacity>=8.2.3  # Retry logic for API calls
backoff>=2.2.1  # Exponential backoff

# Vector databases
faiss-cpu>=1.7.4
pgvector>=0.2.3  # PostgreSQL vector extension client
redis>=5.0.0
pymongo>=4.6.0

# Human-in-the-loop components
databases[postgresql,sqlite]>=0.8.0  # Async database support
asyncpg>=0.28.0  # PostgreSQL async driver
psycopg2-binary>=2.9.7  # PostgreSQL sync driver
slack_sdk>=3.21.3  # Slack integration for notifications
streamlit-extras>=0.3.5  # Enhanced Streamlit components
streamlit-calendar>=1.1.0  # Calendar component for reviewer scheduling
watchdog>=3.0.0  # For file monitoring

# Document processing
pypdf>=3.15.1  # PDF processing
python-docx>=1.0.1  # Word document processing
openpyxl>=3.1.2  # Excel processing
beautifulsoup4>=4.12.0  # HTML parsing
markdown>=3.5.0  # Markdown processing

# Email integration
fastapi-mail>=1.4.1  # Email sending
jinja2>=3.1.2  # Email templates

# Data visualization
plotly==5.18.0
matplotlib==3.8.0
seaborn==0.13.0
altair==5.2.0  # Added for fairness visualizations
scipy==1.12.0  # Added for statistical tests in fairness module

# Model explainability and fairness
shap>=0.42.1  # Model explainability
lime>=0.2.0.1  # Local explainability
fairlearn>=0.9.0  # Fairness metrics
alibi>=0.9.4  # ML explanation and monitoring

# Security
bandit>=1.7.5  # Security linting
safety>=2.3.5  # Dependency security
pyjwt>=2.8.0  # JWT for authentication

# Testing and validation
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
faker>=22.0.0  # Generate fake data
hypothesis>=6.91.0  # Property-based testing

# Kubernetes integration
kubernetes>=28.1.0  # K8s API client

# Observability
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
prometheus-client>=0.17.0
opentelemetry-exporter-prometheus>=1.21.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0

# Security
authlib>=1.2.1
bcrypt>=4.1.0
aioredis>=2.0.1
itsdangerous>=2.1.2

# Monitoring and Observability
grafana-agent>=0.38.0
jaeger-client>=4.9.0  # New: Distributed tracing
datadog>=0.48.0  # New: Production monitoring

# Kubernetes integration
kfp>=2.4.0  # New: Kubeflow Pipelines SDK
seldon-core>=1.16.0  # New: Model deployment on Kubernetes
kubeflow-katib>=0.16.0  # New: Automated ML on Kubernetes

# Cloud Providers - AWS
boto3>=1.34.0
awscli>=1.32.0
amazon-sagemaker-sdk>=2.198.0
aws-cdk-lib>=2.112.0

# Cloud Providers - Azure
azure-storage-blob>=12.19.0
azure-cosmos>=4.5.0
azure-identity>=1.15.0
azure-keyvault>=4.3.0
azure-functions>=1.17.0
azure-mgmt-monitor>=6.0.2
azure-ai-ml>=1.12.0

# Cloud Providers - GCP
google-cloud-storage>=2.13.0
google-cloud-firestore>=2.13.0
google-cloud-kms>=2.21.0
google-cloud-logging>=3.9.0
google-cloud-monitoring>=2.18.0
google-cloud-bigquery>=3.17.0
google-cloud-pubsub>=2.18.0
google-cloud-aiplatform>=1.37.0
google-cloud-container>=2.38.0  # New: GKE integration

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# Testing
pytest-benchmark>=4.0.0  # New: Performance testing

# Web framework
starlette>=0.36.3  # Updated for compatibility with FastAPI
python-multipart>=0.0.6

# Utilities
pydantic-extra-types>=2.3.0
pydantic-settings>=2.1.0
structlog>=24.1.0
python-json-logger>=2.0.7
typer>=0.9.0  # New: CLI application builder
dvc>=3.30.0  # New: Data version control
hydra-core>=1.3.2  # New: Configuration management

# Containerization and orchestration
docker>=7.0.0  # New: Docker SDK for Python
helm>=3.14.0  # New: Helm SDK for Python

# Linting & Type Checking
mypy>=1.8.0
black>=24.1.0
flake8>=7.0.0
isort>=5.13.0
ruff>=0.2.0  # New: Fast Python linter

# LLM and ML Libraries
anthropic>=0.15.0      # Anthropic Claude API
google-generativeai>=0.3.2  # Google Gemini API
sentence-transformers>=2.5.0  # Local embeddings models
cohere>=5.0.0          # Cohere API for embeddings and generation
llama-index>=0.10.0    # Framework for building RAG systems

# LLM Security and Monitoring
langchain-monitor>=0.2.0  # Monitoring for LLM applications
llm-guard>=0.3.0       # Security and filtering for LLM applications
